move optimal circuit to variational result
